+++ src/svlog/hir/visit.rs
@@ -161,7 +161,7 @@ pub fn walk_expr<'a>(visitor: &mut impl Visitor<'a>, expr: &'a Expr, lvalue: boo
                 }
             }
         }
-        ExprKind::Builtin(BuiltinCall::Clog2(arg)) => {
+        ExprKind::Builtin(BuiltinCall::Clog2(arg)) | ExprKind::Builtin(BuiltinCall::Bits(arg)) => {
             visitor.visit_node_with_id(arg, false);
         }
         ExprKind::Ternary(cond, true_expr, false_expr) => {